nengyong

+++ Tools/Include/constants.h
@@ -1,7 +1,7 @@
 #ifndef __OPTION_H_
 #define __OPTION_H_
 
-const int ENEMY_BLUE = 0;
-const int ENEMY_RED = 1;
+const uint8_t ENEMY_BLUE = 0;
+const uint8_t ENEMY_RED = 1;
 
 #endif
